SYNC-1521 Move check
