+++ webs/sync-web/docroot/WEB-INF/src/com/liferay/sync/filter/SyncJSONFilter.java
@@ -14,10 +14,13 @@
 
 package com.liferay.sync.filter;
 
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
+import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.sync.SyncClientMinBuildException;
@@ -29,6 +32,12 @@ import com.liferay.sync.util.SyncUtil;
 
 import java.io.IOException;
 import java.io.OutputStream;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
@@ -169,6 +178,51 @@ public class SyncJSONFilter implements Filter {
 
 	@Override
 	public void init(FilterConfig filterConfig) {
+	}
+
+	public boolean isSyncJSONRequest() {
+		try {
+			String cmd = getParameter(Constants.CMD);
+
+			if (cmd == null) {
+				cmd = StringUtil.read(getInputStream());
+			}
+
+			Object jsonObject = JSONFactoryUtil.looseDeserialize(cmd);
+
+			List<Object> jsonItems = null;
+
+			if (jsonObject instanceof List) {
+				jsonItems = (List<Object>)jsonObject;
+			}
+			else if (jsonObject instanceof Map) {
+				jsonItems = new ArrayList<>(1);
+
+				jsonItems.add(jsonObject);
+			}
+
+			for (Object jsonItem : jsonItems) {
+				Map<String, Map<String, Object>> map =
+					(Map<String, Map<String, Object>>)jsonItem;
+
+				Set<String> keySet = map.keySet();
+
+				Iterator<String> iterator = keySet.iterator();
+
+				String key = iterator.next();
+
+				if (key.startsWith("/sync-web.") ||
+					key.startsWith("/sync-web/")) {
+
+					return true;
+				}
+			}
+		}
+		catch (Exception e) {
+			return false;
+		}
+
+		return false;
 	}
 
 	private static final int _ABSOLUTE_SYNC_CLIENT_MIN_BUILD_ANDROID = 26;
