+++ webs/sync-web/docroot/WEB-INF/src/com/liferay/sync/filter/SyncJSONHttpServletRequestWrapper.java
@@ -14,22 +14,13 @@
 
 package com.liferay.sync.filter;
 
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.StreamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
-
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import javax.servlet.ServletInputStream;
 import javax.servlet.http.HttpServletRequest;
@@ -74,51 +65,6 @@ public class SyncJSONHttpServletRequestWrapper
 			getInputStream());
 
 		return new BufferedReader(inputStreamReader);
-	}
-
-	public boolean isSyncJSONRequest() {
-		try {
-			String cmd = getParameter(Constants.CMD);
-
-			if (cmd == null) {
-				cmd = StringUtil.read(getInputStream());
-			}
-
-			Object jsonObject = JSONFactoryUtil.looseDeserialize(cmd);
-
-			List<Object> jsonItems = null;
-
-			if (jsonObject instanceof List) {
-				jsonItems = (List<Object>)jsonObject;
-			}
-			else if (jsonObject instanceof Map) {
-				jsonItems = new ArrayList<>(1);
-
-				jsonItems.add(jsonObject);
-			}
-
-			for (Object jsonItem : jsonItems) {
-				Map<String, Map<String, Object>> map =
-					(Map<String, Map<String, Object>>)jsonItem;
-
-				Set<String> keySet = map.keySet();
-
-				Iterator<String> iterator = keySet.iterator();
-
-				String key = iterator.next();
-
-				if (key.startsWith("/sync-web.") ||
-					key.startsWith("/sync-web/")) {
-
-					return true;
-				}
-			}
-		}
-		catch (Exception e) {
-			return false;
-		}
-
-		return false;
 	}
 
 	private ByteArrayOutputStream _byteArrayOutputStream;
